Move ant colony related files to the same folder.
