+++ src/Qubic.vcxproj.filters
@@ -188,17 +188,17 @@
     <ClInclude Include="mining\mining.h">
       <Filter>mining</Filter>
     </ClInclude>
-    <ClInclude Include="mining\ant_colony.h">
-      <Filter>mining</Filter>
+    <ClInclude Include="mining\ant_colony\ant_colony.h">
+      <Filter>mining\ant_colony</Filter>
     </ClInclude>
-    <ClInclude Include="mining\ant_pending_solutions.h">
-      <Filter>mining</Filter>
+    <ClInclude Include="mining\ant_colony\ant_pending_solutions.h">
+      <Filter>mining\ant_colony</Filter>
     </ClInclude>
-    <ClInclude Include="mining\ant_colony_bpp9000.h">
-      <Filter>mining</Filter>
+    <ClInclude Include="mining\ant_colony\ant_colony_bpp9000.h">
+      <Filter>mining\ant_colony</Filter>
     </ClInclude>
-    <ClInclude Include="mining\ant_colony_snapshot.h">
-      <Filter>mining</Filter>
+    <ClInclude Include="mining\ant_colony\ant_colony_snapshot.h">
+      <Filter>mining\ant_colony</Filter>
     </ClInclude>
     <ClInclude Include="mining\trit_pack.h">
       <Filter>mining</Filter>
@@ -473,6 +473,9 @@
     </Filter>
     <Filter Include="mining">
       <UniqueIdentifier>{df525479-7504-470c-a25a-de4af8be0e5d}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="mining\ant_colony">
+      <UniqueIdentifier>{7b1f3a52-9c4e-4d2a-b8e6-2a5c9d417f60}</UniqueIdentifier>
     </Filter>
     <Filter Include="logging">
       <UniqueIdentifier>{d334594b-f24d-440e-949a-c791aa13f867}</UniqueIdentifier>
