+++ src/mining/ant_colony/ant_colony_bpp9000.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "ant_colony.h"
-#include "../score.h"
+#include "mining/ant_colony/ant_colony.h"
+#include "score.h"
 
 // Binds the colony to bpp9000. This is the only place a concrete scorer is named, which is why
 // ant_colony.h itself can stay free of score.h and everything it drags in.
